Adding the diagnosticCriteria values in the WHOFIC export.
Note that only the first value is exported, to be in line with the
expected single value requirement in the new WHOFIC model.
Where there more than one diagnosticCriteria values exist, they should
be replaced by a single value that contains the properly merged version
of those values, before the final migration.
Also the prefix "!markdown" is added to the beginning of the value so
that the already existing Markdown values are recognized as such, and be
rendered properly, in iCAT-X.